PID implemented, working with 45 pwm

+++ pwm.h
@@ -1,7 +1,7 @@
 #ifndef PWM_H_
 #define PWM_H_
 
-void SetDutyCycle(unsigned int leftDuty, unsigned int rightDuty, unsigned int Frequency, int dir);
+void SetDutyCycle(float leftDuty, float rightDuty, unsigned int Frequency, int dir);
 void InitPWM();
 void PWM_ISR();
 void SetServoDutyCycle(float DutyCycle, unsigned int Frequency, int dir);
